test: move issue 66827 regression into existing suite
